added parity functionality, makefile for compilation and gitignore file

+++ major1.h
@@ -1,5 +1,6 @@
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdbool.h>
 
 
-void parity(int value);//parity function
+void parity(unsigned long int value);//parity function
